cnn/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
werkzeug==3.1.0
tensorflow-cpu==2.16.1
pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
h5py==3.11.0
